+++ src/lib/actions/anilist.ts
@@ -1,17 +1,20 @@
-import { NextResponse } from 'next/server'
+'use server'
+
+import type { AnimeResponse } from '@/schemas/anime'
 import { z } from 'zod'
 import { AnimeResponseSchema } from '@/schemas'
 
-export async function GET(req: Request) {
-  try {
-    const { searchParams } = new URL(req.url)
-    const userName = searchParams.get('userName')?.trim() ?? ''
-
-    if (userName.length <= 1) {
-      return NextResponse.json({ error: 'Username Issue' }, { status: 400 })
+export const fetchAnilistData = async (
+  userName: string,
+): Promise<ActionResponse<AnimeResponse>> => {
+  if (userName.length <= 1) {
+    return {
+      success: false,
+      message: 'Username Issue',
     }
+  }
 
-    const query = `
+  const query = `
       query ($userName: String) {
         MediaListCollection(userName: $userName, type: ANIME) {
           lists {
@@ -45,8 +48,9 @@ export async function GET(req: Request) {
             }
           }
         }
-      }    `
+      }`
 
+  try {
     const response = await fetch('https://graphql.anilist.co', {
       method: 'POST',
       headers: {
@@ -59,20 +63,33 @@ export async function GET(req: Request) {
     if (!response.ok) {
       const errorBody = await response.text()
       console.error('GraphQL API error:', errorBody)
-      return NextResponse.json({ error: 'Upstream API error' }, { status: 502 })
+      return {
+        success: false,
+        message: 'Upstream API error',
+      }
     }
 
     const data = await response.json() as unknown
 
     const parsedResponse = AnimeResponseSchema.parse(data)
-
-    return NextResponse.json(parsedResponse)
+    return {
+      success: true,
+      message: 'Data fetched successfully',
+      data: parsedResponse,
+    }
   }
   catch (error) {
     if (error instanceof z.ZodError) {
       console.error('Zod validation error:', error.errors)
-      return NextResponse.json({ error: error.errors }, { status: 400 })
+      return {
+        success: false,
+        message: error.errors.map(e => e.message).join(', '),
+      }
     }
-    return NextResponse.json({ error: 'Internal Server Error' }, { status: 500 })
+    console.error('Internal Server Error:', error)
+    return {
+      success: false,
+      message: 'Internal Server Error',
+    }
   }
 }